+++ graphic/ViewEternityForMPJE.java
@@ -48,7 +48,7 @@ public class ViewEternityForMPJE extends JFrame implements KeyListener {
 	private boolean pauseAll = false;
 	private boolean pauseGraphic = false;
 	private long refresh_milis;
-	private int prevCount = 0;
+	private long prevCount = 0;
 	private int innerCount = 1; // este controla el refresco de las piezas por segundo
 	private static final int COUNT_SPEED = 5; // factor de velocidad de refresco del contador de piezas
 	private StringBuilder titleRefreshed = new StringBuilder(128);
@@ -140,7 +140,7 @@ public class ViewEternityForMPJE extends JFrame implements KeyListener {
     	
     	// calculo piezas (o nodos) por segundo. Se actualiza cada 1/COUNT_SPEED segs
     	if (innerCount >= frequency) {
-    		int currentCount = 0;
+    		long currentCount = 0;
 			currentCount = SolverFasterMPJE.count_cicles - prevCount;
 			prevCount = SolverFasterMPJE.count_cicles;
